fix: move DNSMessage into message.py
