+++ src/zeroconf/_protocol/message.pxd
